+++ tests/testthat/test_spss_swap.R
@@ -1,4 +1,10 @@
+# Testing the Input -------------------------------------------------------
 test_that("spss_swap stops with an error for invalid argument types", {
-  # Creating a test flextable
   expect_error(spss_swap(mtcars), "Invalid argument type")
+  expect_error(spss_swap(),"Need to specify the mandatory argument")
 })
+
+
+# Testing Output ----------------------------------------------------------
+# to be written
+